+++ src/services/utils.ts
@@ -136,4 +136,5 @@ export default {
   getLocalForageItem,
   setLocalForageItem,
   removeLocalForageItem,
+  getLocalForageKeys,
 }